+++ src/app/ja/writing/page.tsx
@@ -1,7 +1,7 @@
 import WritingPageContent from '@/components/containers/pages/WritingPage'
 import { loadBlogPosts } from '@/libs/dataSources/blogs'
-import { MicroCMSAPI } from '@/lib/microCMS/apis'
-import { createMicroCMSClient } from '@/lib/microCMS/client'
+import { MicroCMSAPI } from '@/libs/microCMS/apis'
+import { createMicroCMSClient } from '@/libs/microCMS/client'
 
 export const metadata = {
   title: 'Writing',
