+++ src/app/work/[slug]/page.tsx
@@ -1,5 +1,5 @@
-import { MicroCMSAPI } from '@/lib/microCMS/apis'
-import { createMicroCMSClient } from '@/lib/microCMS/client'
+import { MicroCMSAPI } from '@/libs/microCMS/apis'
+import { createMicroCMSClient } from '@/libs/microCMS/client'
 import Container from '@/components/tailwindui/Container'
 import SimpleLayout from '@/components/tailwindui/SimpleLayout'
