+++ src/components/containers/FeaturedContent.tsx
@@ -2,10 +2,10 @@ import Link from 'next/link'
 import Image from 'next/image'
 import Container from '@/components/tailwindui/Container'
 import { loadBlogPosts } from '@/libs/dataSources/blogs'
-import { MicroCMSAPI } from '@/lib/microCMS/apis'
-import { createMicroCMSClient } from '@/lib/microCMS/client'
+import { MicroCMSAPI } from '@/libs/microCMS/apis'
+import { createMicroCMSClient } from '@/libs/microCMS/client'
 import type { FeedItem } from '@/libs/dataSources/types'
-import type { MicroCMSProjectsRecord, MicroCMSPostsRecord, MicroCMSEventsRecord } from '@/lib/microCMS/types'
+import type { MicroCMSProjectsRecord, MicroCMSPostsRecord, MicroCMSEventsRecord } from '@/libs/microCMS/types'
 import DateDisplay from '@/components/ui/DateDisplay'
 import Tag from '@/components/ui/Tag'
 import BackgroundDecoration from '@/components/ui/BackgroundDecoration'
